feat/add-responsible-court-to-result-page: use isErfolgAnalog for stringReplacement
