+++ app/domains/fluggastrechte/vorabcheck/__test__/stringReplacaments.test.ts
@@ -15,9 +15,10 @@ import {
   getButtonURLForClaimViaPost,
 } from "~/domains/fluggastrechte/vorabcheck/stringReplacements";
 import { getRouteCompensationBetweenAirports } from "~/services/airports/getRouteCompensationBetweenAirports";
-import { hasAirportPartnerCourt } from "~/services/airports/hasPartnerCourt";
+import { isErfolgAnalog } from "../services/isErfolgAnalog";
 
 vi.mock("~/services/airports/getRouteCompensationBetweenAirports");
+vi.mock("../services/isErfolgAnalog");
 
 const mockedGetRouteCompensationBetweenAirports = vi.mocked(
   getRouteCompensationBetweenAirports,
@@ -279,20 +280,19 @@ describe("hasCompensationShortDistance", () => {
   });
 
   describe("getButtonURLForClaimViaPost", () => {
-    vi.mock("~/services/airports/hasPartnerCourt");
-    it("should return the link to ergebnis/erfolg when partnerCourt exists", () => {
-      vi.mocked(hasAirportPartnerCourt).mockReturnValue(true);
-      const actual = getButtonURLForClaimViaPost({});
-      expect(actual).toStrictEqual({
-        claimViaPostButtonURL: "/fluggastrechte/vorabcheck/ergebnis/erfolg",
-      });
-    });
-    it("should return the link to ergebnis/erfolg-analog when there is no partnerCourt", () => {
-      vi.mocked(hasAirportPartnerCourt).mockReturnValue(false);
+    it("should return the link to ergebnis/erfolg-analog when isErfolgAnalog is true", () => {
+      vi.mocked(isErfolgAnalog).mockReturnValue(true);
       const actual = getButtonURLForClaimViaPost({});
       expect(actual).toStrictEqual({
         claimViaPostButtonURL:
           "/fluggastrechte/vorabcheck/ergebnis/erfolg-analog",
+      });
+    });
+    it("should return the link to ergebnis/erfolg when IsErfolgAnalog is false", () => {
+      vi.mocked(isErfolgAnalog).mockReturnValue(false);
+      const actual = getButtonURLForClaimViaPost({});
+      expect(actual).toStrictEqual({
+        claimViaPostButtonURL: "/fluggastrechte/vorabcheck/ergebnis/erfolg",
       });
     });
   });
